author tilføjet check af bookswritten

+++ models/authors.js
@@ -1,5 +1,6 @@
 const mongoCon = require('../dbs')
 const { getNextId } = require('../lib/getNextId')
+const { hasBoksWritten } = require('../lib/authorExists')
 const msg = require('../lib/messages')
 const authorsCollection = 'authors'
 
@@ -51,6 +52,9 @@ module.exports = {
 
   async deleteById (author_id) {
     try {
+      const booksWritten = await hasBoksWritten(author_id)
+      if (booksWritten) return msg.author_has_books(author_id)
+
       let db = await mongoCon.getConnection()
       const result = await db.collection(authorsCollection).findOneAndDelete({ id: author_id })
       return (result.ok === 1 && result.value) ? msg.record_deleted(author_id,'Author') : msg.record_not_found(author_id, 'Author')
